sample-cdk/requirements.txt
aws-cdk-lib==2.110.1
constructs>=10.0.0
pytest==7.4.3
boto3==1.34.28